+++ src/Circuit/Parser/Csv/Lexer.hs
@@ -1,5 +1,4 @@
 {-# LANGUAGE BangPatterns #-}
-{-# LANGUAGE DeriveGeneric #-}
 
 -- | Fast imperative CSV tokenizer over strict 'ByteString'.
 --
@@ -16,12 +15,10 @@ module Circuit.Parser.Csv.Lexer
   )
 where
 
-import Control.DeepSeq (NFData)
 import Data.ByteString (ByteString)
 import Data.ByteString qualified as BS
 import Data.ByteString.Unsafe qualified as BSU
 import Data.Word (Word8)
-import GHC.Generics (Generic)
 
 -- $setup
 -- >>> import Data.ByteString.Char8 qualified as C
@@ -31,9 +28,7 @@ data CsvToken
   = CField ByteString
   | CQuoted ByteString
   | CRowEnd
-  deriving (Eq, Show, Generic)
-
-instance NFData CsvToken
+  deriving (Eq, Show)
 
 -- | Tokenize a strict 'ByteString' as CSV.
 --